requirements.txt
# Web框架
Django==5.0.2
djangorestframework==3.14.0

# 环境配置
python-dotenv==1.0.1

# 区块链相关
web3==6.15.1
eth-account==0.10.0
solana==0.30.2
solders==0.18.1
mnemonic==0.20

# 加密和数据库
cryptography==42.0.5
SQLAlchemy==2.0.25
base58==2.1.1

# 网络请求
requests==2.28.1

# 其他依赖
PyMySQL==1.1.0
PyNaCl==1.5.0
Pillow==10.2.0

# ... 